Search of entity in the initial question
Subject in a text format is to be found in the initial question for adding START and END tokens outside of it. The process of finding text of the subject in the initial question was organized as follows:

1) lower text and find exact matching
2) lower text and find the main part of the candidate without the part in brackets
3) lower text and try to find it using fuzzywuzzy package (count for typos, misprints, etc)

After all steps (1,2,3) auto_ner function checks the matching for being adequate (if the length of the found entity is less then 2, then matching is wrong).

In the end for all sentences where there is no matching, “start” token is set at the beginning of the sentence while “end” token in the end.

As a result, you obtain SImpleQuestions dataset where subject is marked with START and END tokens.