Spacy NER finetuned on SimpleQuestions dataset
The file takes SimpleQuestions dataset, converts it to the format suitable for Spacy NER training and train the model. The trained model is used for NER prediction and adding of the START and END tokens to the initial dataset.

and

Merging commits for all renaming (renamed branch with
experiments/ner_for_mgenre; renamed ipynb notebooks with issue numbers
in the title; created folder in my branch for experiemnts) and
resubmitted the class for NER insertion with black and pylint updates
for the style code.